Read currency values as decimal instead of float (#280)
DbfValueCurrency was backed by float and divided the raw scaled
integer by 10000.0f, while the column schema advertised
typeof(decimal) for 'Y' columns. GetDecimal() therefore threw
InvalidCastException on currency columns, and large amounts lost
precision to float rounding.

Currency values are stored as a 64-bit integer scaled by 10,000, so
dividing by 10000m yields the exact value. ToString() output is
unchanged (decimal division produces minimal scale).

BREAKING: code calling GetFloat()/GetValue<float>() on currency
columns must use GetDecimal()/GetValue<decimal>() instead, matching
the advertised column schema.

+++ src/DbfDataReader/DbfValueCurrency.cs
@@ -3,7 +3,7 @@ using System.Globalization;
 
 namespace DbfDataReader
 {
-    public class DbfValueCurrency : DbfValue<float?>
+    public class DbfValueCurrency : DbfValue<decimal?>
     {
         public DbfValueCurrency(int start, int length, int decimalCount) : base(start, length)
         {
@@ -14,8 +14,9 @@ namespace DbfDataReader
 
         public override void Read(ReadOnlySpan<byte> bytes)
         {
+            // currency is stored as a 64-bit integer scaled by 10,000
             var value = BitConverter.ToInt64(bytes);
-            Value = value / 10000.0f;
+            Value = value / 10000m;
         }
 
         public override string ToString()
